OpenStack: remove jobs from release-controller
The idea here is to reduce the number of jobs hitting our CI cloud as
it is performing particularly bad under load, to allow more important
jobs (say 4.4) to run better.

This is only a temporary measure and should be reverted as soon as the
number of new images per day drops significantly, or even before when
our CI is able to accomodate the load.