Setup openshift-metal3/dev-scripts for Prow
dev-scripts is the automation used by baremetal IPI developers to create
a virtualized baremetal environment on a single server. This is used by
the e2e-metal-ipi test jobs to create an environment for installing OCP
on a single Packet host.

This first PR just enables some plugins on the repo to make sure Prow is
working.