Setup openshift-metal3/dev-scripts for OpenShift CI
dev-scripts is the automation used by baremetal IPI developers to create
a virtualized baremetal environment on a single server. This is used by
the e2e-metal-ipi test jobs to create an environment for installing OCP
on a single Packet host.

This PR enables CI testing on dev-scripts itself.  `dev-scripts` master
supports 4.4 and later, so this adds test jobs to test both 4.4 and 4.5
against each PR.