oVirt: e2e, remove unnececery image-registry call
In the past we did some operations on the cluster
after the installation and checked that we can get
the image-registry from the cluster.
This is not needed now and since we have set -e
if the command will fail then the script will
fail 11d that is not what we want, if the cluster
is installed but has a problem we want the test pod
to fail.

Signed-off-by: Gal Zaidman <gzaidman@redhat.com>