Step registry: Deploy loki and gather all container logs
Following https://steps.svc.ci.openshift.org/help and based on already existing steps.
The idea is to deploy loki which is resource low-cost solution for collecting container logs.
Given the current gather step collects only container logs of existing pods (with the last
previous as well), we can't see the entire history of what happend. E.g. in case a container
is restarted more than once.

The PR defines two new steps:
1. loki deployment (with inlined manifests)
2. querying loki and collecting all the container logs

Each query for a container log has timeout set to 1 minute.
In case it's not sufficient (for container with many log lines), the timeout
needs to be increased.

Also, the PR is part of an iterative effort. The idea is to deploy loki,
collect the logs and see if the current implementation is sufficient.
Improve the script accordingly if it is not.

Updating kube-apiserver-operator, kube-controller-manager-operator, kube-scheduler-operator and origin e2e-aws tests first.