capture output as-is
Make this step less fragile by simply capturing the must-gather output as-is and not trying to interpret it. 

Currently, this task breaks in 4.1 as the task assumed the structure introduced in 4.2. 

I've also removed the explicit specification of an image, allowing the underlying must-gather command to determine the version specific default.