baremetal: bump to c3.medium and 2 workers
We're seeing a lot of flakes in CI, and there's been some suggestion
that our deployment of 3 control plane nodes and 1 worker isn't a
supported configuration and may be some source of the flakes.

The idea is to try this configuration for a while and see if the
stability of the periodics improve.