ci-operator/step-registry/origin/e2e/test: Drop --insecure-skip-tls-verify
The --insecure-skip-tls-verify tradition dates back to e071153af0
(don't validate TLS when collecting from CI, 2019-02-25, #2950).  But
the motivation there was "even if we flub certs, we still want to
collect logs".  Talking to the cluster during the *test* run is not
about log collection, and we definitely want to fail tests if we flub
the certs.  I expect this line got its --insecure-skip-tls-verify
accidentally in e071153af0.