enable the topology manager single-numa-node conf
Add a new CLUSTER_VARIANT to configure the kubelet topology manager
with the single-numa-node policy, in order to make it possible
to run the e2e tests.

This is because this setting is the only one which allows a meaningful
testing, because guarantees unambiguous results. Thus, this setting
is a precondition to have the topology manager e2e tests run.

Signed-off-by: Francesco Romani <fromani@redhat.com>