oVirt e2e: increase test pod cpus and change mem resources
When we log into the test pod when running openshift-tests
we see very high cpu usage and load average of ~100-200.
Latly we have been seeing timeouts on test pod [1].
Added a workaround on [2] which is tracked for removel on [3].
On our tests we sow that increasing the CPUs on test pod solved
the timeout problem.

[1] https://ci-search-ci-search-next.svc.ci.openshift.org/?search=infrastructures%2Fcluster%3A+dial+tcp%3A+i%2Fo+timeout&maxAge=168h&context=2&type=bug%2Bjunit
[2] https://github.com/openshift/release/pull/9279
[3] https://bugzilla.redhat.com/show_bug.cgi?id=1839463

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>