spyglass: only highlight 'error:' when it appears at the beginning of the line
Ansible reports 'error:' in messages even when it's not a problem. For
example, a task in a play may be checking whether or not a thing exists,
and failing could be a normal operation. When it does this though,
it's reported in json.

The only 'error:' we should highlight is one output by `oc`, which
always start at the beginning of the line.