baremetal: collect additional logs
This adds collection of the libvirt logs, which with the latest
metal3-dev-env changes includes the serial consoles of the VM's. This
allows better troubleshooting of PXE failures.

This also ensures `$HOSTNAME` is included in each tarball's filename,
which includes the CI namespace. This helps indicate where a particular
log archive came from.