alerts: add an alert for ci-operator errors
After observing what outages look like in this metric, we can now add an
alert for them. This is a very liberal definition as it sets a hard
lower bound on frequency and does not filter out for reasons we know are
inside of user code. However, from observation of the data, error rates
only spike above this value if there is some infrastructural issue. We
surely will need to tune this in the future, but it's a good start.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>