oVirt e2e: skip minimal tests on PR jobs
In our CI PRs we are overcommiting resources to our nodes to be able to support a number of PRs/release jobs in parallel.
Reacently the test case:
"Prometheus when installed on the cluster shouldn't report any alerts in firing state apart from Watchdog and AlertmanagerReceiversNotConfigured"
started catching alert "KubeMemoryOvercommit" as firing and failing PR jobs,
Since the run-minimal-tests custom test suite contained only 2 test cases, and the mainly checks that the custer is really up
we decided to skip the minimal test suite for PRs only and rely on curl to the API server.

This is taken from openstack approch on [1]

[1] https://github.com/openshift/release/blob/master/ci-operator/templates/openshift/installer/cluster-launch-installer-openstack-e2e.yaml#L238-L243

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>