ci-operator/step-registry/openshift: Push e2e flavors under e2e/
This collects similar workflow definitions together instead of having
a growing dump of sibling directories.

Also use symlinks to parent OWNERS files for subdirectories had the
same content.  I haven't used symlinks for things like:

* aws/ovn/OWNERS
* vsphere/ovn/OWNERS

despite them having the same content, because it's not clear which
location would be more canonical.  Will be nice if/when [1] gets fixed
and we can go back to using aliases.

[1]: https://github.com/openshift/ci-tools/pull/115#discussion_r350000636