config: promote from 4.5
The config-brancher invocation used to cut over to the new branch was
not fed the right flags. As a result, the promotion field was not
reverted correctly. While we can "bump" master back to 4.5 and then bump
back to 4.6 correctly, that will also carry forward changes on the
master config into the 4.5 configs that happened after branch-cut, like
Go version changes, etc. So, the simplest way is sed.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>