config: update replace names to match builder
Users often do two types of builds in CI: first, a simple build using
the `build_root` in order to create images with which they can run
static analysis, etc. Second, a build that uses the ART-provided builder
image in a multi-stage process that mimics OSBS and gives them feedback
on whether their builds will function in the productized pipeline. We
need to make sure that these two match.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>