steps: allow install, test, gather and destroy to run off of a jump host
Currently the CI pod that runs the installer/openshift-test/oc(gathering) needs access to the internet for various things.

-    the cluster that was created, kube api and the openshift-ingress
-    the cloud APIs
-    DNS resolution of both types of endpoints mentioned above.

The first one is about getting the CI pod access to internal network ips, the second one is not necessary for internal clusters specifically but for certain testing like AWS GovCloud emulation testing, we need to make sure we are originating all the request from inside the network itself to use PrivateLink service endpoints and the last one is important because the internal cluster does not publish the dns names to Internet, the dns resolution of api.clusterdomain for example is only possible from inside the cluster and also affects the usage of PrivateLink service endpoints.

So I started with 3 idea and tried to implement those using what we have.

1. So started looking into complete transparent VPNs like [sshuttle| https://github.com/sshuttle/sshuttle] or OpenVPN
Using these should allow all the traffic to be force through the VPN into the internal network and also provide dns resolution.
But when trying to use them in container esp on an OpenShift clsuter,

-    Both of them require sudo/root priviledges + NET_CAP like capabilities in the pod. With openshift this is very difficult to achieve as far as read.
-    Also these require modyging network devices and I could not find a way to even test this on OpenShift clsuter.

2. Then I started looking into, ssh+rsync
For a given step we take the commands.sh and use rsync to copy that plus any binaries we need to the bastian-host's temporary workspace. Plus we also sync all the shared_dir, artifacts_dir and any other environment variables to ensure we can run the step properly.
This is approach i have taken in commit but there are a lot of downsides to this approach,

-    Requires a lot of wrapping and needs an alternate steps with jump handling for each step today. https://issues.redhat.com/browse/DPTP-1259 can help reduce one type of dupliation but there is still which binaries to copy, what env variables should we tranfer and what if there are other files that need to be copied.
-    The shared_dir, artifacts_dir syncing is a little brittle, the impl should be good enough but there is alittle more to be desired in terms of retries etc.
-    The test require 3 cpus, 4 gb of ram and that means we can't shared bastians and we might have to bring one up per CI job

3. The last idea i have is using SOCKS5 proxy + pod dnspolicy forced to dns resolver of the internal network.

Since most of our binaries are Go binaries. Go support HTTP_PROXY="socks5://<>" and this would allow us to access the internal IPs over the proxy as long as we wrap the commands.sh with a local ssh proxy.

The next requirement is to force the DNS to that of the internal network. K8s pods allow defining dnsPolicy and nameservers for the pods.
```yaml
dnsPolicy: None
dnsConfig:
  nameservers:
  - <the chosen one, probably needs to be dynamic per the steps>
```
This should technicaly allow the pod to use bastian dns resolvers already setup to allow dns resolution of internal services of our pre-existing networks. Now this requires that ci-operator can read these values from like SHARED_DIR and then setup the dns for the pod running the step accordingly.
There is one down side, no containers in that pod will have access to the CI cluster's services so current implementation of shared-secrets-copier might have to be updated.