build01: add openshift-dev key
It's pretty hard to debug the runtime on a node if one can't replace the runtime. Unfortunately, `oc debug node/` only works by dropping a pod on the system, and thus the CRI-O binary can't be replaced to debug.

Instead, we should add ssh keys (just for nodes with the debug role) to be able to access nodes directly

Signed-off-by: Peter Hunt <pehunt@redhat.com>