Re-Enable promotionreconciler
I've played around with it and can not find an explanation for the high
token usage we saw. On the first attempt it used ~50 tokens and after
another restart it used five. I've played around with all of the five
via curl and the etags are set correctly (getting it multiple times
returns the same etag).

Hence we should re-enable it, hope it won't break anything, wait until
we see high token usage in the metrics and then look at the logs for
that period.