oVirt: add env vars to teardown pod
On cluster destroy we check if the value is set.
Due to a bug on the installer template destory condition this was never an issue.
Now that it is fixed we want to avoid from removing the template.
So we export the env variable.

Exporting an empty string to not breack the CI untill the installer PR will be merged.

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>