CI: Fix cluster password leak in the installer logs
When the cluster installation succeeds and the teardown fails,
the installer logs show the username and the password of the cluster
created in the log files which the installer logs during the create
cluster and the wait-for install-complete commands.

Added a fix that would find the log file and replace the password string
with "REDACTED". This fix is done exactly after every create cluster command(mostly IPI)
and the wait-for install-complete(mostly UPI) commands are done.