Retry e2e-metal-ipi for release informers
This sets maxRetries = 3 for e2e-metal-ipi. The main reason for wanting
to retry is that reliability for Packet has been an issue, we're failing
some builds at busy times because of inability to obtain a packet host.
Failures vary, but a few common reasons come up. For example, we rely on
'any' to route us to the best Packet datacenter, but it doesn't seem to
consider instance type availability.