Add a new base image for UBI 8 and RHEL 8 content
We will be assessing whether issues arise on RHEL7 workers as a
result of moving all of our images to UBI8 across the payload. To
do that, create a new base-8 and base-7 tag for the two types, and
continue to publish "base" as the default. This will allow us to
test individual projects.

Correct an issue with jobs pointing to the wrong release component.