ci-operator/jobs/openshift/cluster-version-operator: Move 4.4 and later presubmits to GCP
AWS is currently at CI-capacity limits, so move over to GCP to get
more elbow room.  Generated by adjusting the config/ entry by hand and
then running:

  $ make update

I'd initially just shifted master, because the bulk of our CI is on
master-targeted pulls.  But Scott asked for 4.4 and later [1], so
that's what I'm doing now.

I'd like to move to generic names like "e2e" and "e2e-upgrade",
because we don't care (from the CVO side) what platform these run on.
Generic names will mean smaller diffs if we pivot the workflow and
cluster_profile to a different provider in the future.  But Scott is
not (yet ;) ready to break with tradition here [2], so punting on that
for now.

I've dropped "e2e-aws-upi", because the CVO shouldn't care about that
level of provisioning detail.  We just need "a cluster".  We've had
the optional UPI job since 9f61634fe5 (cluster-version-operator: add
aws-upi optional job, 2020-02-11, #7125), where it was motivated [3]
by a desire to test the cluster_installer metric [4].  Now that that
metric is well-established, we can drop the option.  And we can always
regrow it if we need it again later.

[1]: https://github.com/openshift/release/pull/10046#issuecomment-654910661
[2]: https://github.com/openshift/release/pull/10046#issuecomment-654519730
[3]: https://github.com/openshift/release/pull/7125#issue-373972653
[4]: https://github.com/openshift/cluster-version-operator/pull/319