app.ci: Add release signer to app.ci
The signer will create a signature in the GCS storage bucket
openshift-ci-release at /releases/signatures/* that OKD and CI
clusters can validate against. This will enable OKD to upgrade
from version to version.

The bucket this publishes to is separate from the official OCP
releases to provide security isolation and prevent denial of
service attacks for upgrades by deleting all signatures. The
publisher account only has permission to create objects.

The signer key for openshift-ci is stored in bitwarden and should
only be deployed to the app.ci cluster. If compromised, we would
have to create a new key, resign everything, and then instruct all
OKD users to force upgrade. In the future we may add a second
signing key we keep hidden for future rotation if necessary.

The release controller in audit mode will only sign releases that
pass `oc adm release info --verify`, which checks integrity and
availability, but not reproducibility. Future verification checks
may be more extensive and combine other verification.