Include all files in cluster profile CMs
This is IMO easier resolution of our failure to rehearse cluster profile
changes, which currently fail on validation because various non-YAML
(.gitignore, .type, OWNERS) files are not included in the CM, and
pj-rehearse has a built-in validation that all files are covered by some
config-updater pattern. That validation is there to prevent rehearsing
jobs using templates that are already not covered by config-updater, and
as a side effect, this validation broke cluster profile rehearsals.

Alternatively, we could be smart in pj-rehearse and only make sure that
*.yaml files are covered, or something similar, but this change achieves
the same thing for a very little cost.