ci-operator/config/openshift/cluster-version-operator: Generic e2e-gcp -> e2e for 4.4+
In f5320cde00 (ci-operator/jobs/openshift/cluster-version-operator:
Move 4.4 and later presubmits to GCP, 2020-07-02, #10046), I'd
mentioned generic names as a way for component teams to say "we don't
care which platform, other folks can pick whatever they want to
balance platform volume vs. capacity".  Scott pushed back based on
lack of precedent [1,2].  And we currently do rely on platforms in
job-names for monitored success rates.  But these are presubmits, and
presubmits are noisy (e.g. if you pull-request some buggy code, your
presubmits will fail, but the delivered product is not affected by
your in-flight bugs).  So in the presubmit case we are not constrained
by job-name-based health reporting.  And we run a lot of presubmit
volume, so even if this approach only works in the presubmit case, it
gives us a lot of rebalancing flexibility.

Benefits for component teams (the CVO maintainers in the case of this
commit) include not having to care about where their jobs get
scheduled.  And with the job names not changing during rebalances,
they don't have to remember to /skip or /refresh or whatever, because
Prow would understand that it's the same effective test regardless of
the underlying platform.

Benefits to rebalancing admins is that they don't have to ask
component teams to opt-in at rebalance time.  Teams can opt in with
this pattern, and rebalancing admins can just search for e2e steps
that do not include a platform slug in their 'as' config name.

Generated by manually changing ci-operator/config/... and then running:

  $ make update

[1]: https://github.com/openshift/release/pull/10046#issuecomment-654519730
[2]: https://github.com/openshift/release/pull/10046#issuecomment-654911180