ci-operator/step-registry/openshift/kubernetes/e2e/aws: Include "aws" in workflow name
It's an AWS workflow, and we might grow additional workflows for other
platforms in the future.  Including the platform in the workflow name
makes space for that, and sets the stage for automated tooling that
can rebalance between platforms by shifting load from
openshift-kubernetes-e2e-aws to openshift-kubernetes-e2e-gcp, or
whatever.