Allow merging in openshift/kubernetes with invalid owners label
Maintaining anything other than the root OWNERS file for
openshift/kubernetes is likely to cause merge conflicts for carries,
so non-root OWNERS files will be left untouched. Since these files
will contain many upstream users that are not members of the openshift
org, the invalid owners plugin will label all openshift/kubernetes PRs
with `do-not-merge/invalid-owners-file` which is intended to block
merges. Since there's no easy way to disable the plugin, the next best
thing is ensuring that merging is allowed when the
`do-not-merge/invalid-owners-file` label is present.