teardown: Separate contents of pod log directory into subfolders
Create a subfolder per namespace and add the node name to the end
of the filename. This allows better organization of that directory
and makes it easier to correlate pod behavior across nodes. This
simplifies lookups when trying to correlate a pod with the other
infrastructure pods on the same node.