add ports for additional ppc64le libvirt ci system
This allows us to split jobs onto two systems. No new ports added, only new
services. There shouldn't be conflicts since the clusters will run on seperate
systems (as set by the hostname in the job definition). It would make sense to
keep increasing port numbers to be in line with the same numbering schema, but
we'd hit the end of the port range.

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>