ci-operator/step-registry/openshift/e2e/fips: Add FIPS steps
And use them for master installer CI.

The implementation for the openshift-e2e-fips step is something of a
union of the old-style
ci-operator/templates/openshift/installer/cluster-launch-installer-e2e.yaml
and the gather-aws-console step, which also iterates over nodes.  I'm
using 'read' instead of the {1..5} to iterate, because there may not
be exactly five nodes.

The openshift-e2e-aws-fips-default worklow demonstrates that the
default install does not enable FIPS.  I'm just using that for the
installer as well; it seemed a bit heavy to add to openshift-e2e-aws.

The ci-operator/jobs changes were generated with:

  $ make update