OpenStack: use ephemeral disks for control plane
While looking at ways to improve storage performances for control plane
nodes in order to satisfy etcd requirements we started using
`performance` volume type in MOC, however latency was still too high so
we switched to mounting etcd partition in a ramdisk. There was no longer
a need to boot VMs from volumes but we chose to keep it around to
increase coverage.

Lately MOC started to perform poorly. We want to switch back to
ephemeral disks and see if it makes a difference, by reducing the load
on ceph storage.