add optional e2e-aws-upgrade job
"Application behind service load balancer with PDB is not disrupted"

The upgrade test suite has a test that uses the cloud
provider load balancer, but only has a job for master that runs
in gcp.

In order to discard problems with cloud provider load balancers, we
can run the test in aws to compare the results.
The job is optional and only can be run on demand.

Signed-off-by: Antonio Ojea <aojea@redhat.com>