release: dereference environment variables
The `"$(name)"` form does not play well with YAML as it, itself, does
not need quotations to be valid. However, the branch is a float, so once
it is dereferenced the output does need quotes to be a string.
Round-tripping serialization is basically impossible with this caveat,
and we expect these jobs to be autogenerated soon, so we can dereference
these variables and not incur a maintenance cost.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>