OpenStack: gather nova logs for all provisioned servers
The previous implementation wasn't that useful as it would only gather
nova console logs for servers that joined the cluster. We want to get
the logs for all provisioned instances, in order to help troubleshoot
when they didn't join the cluster.

This moves the nova logs from the nodes/node_name directory to the
nodes/ directory, so we can know at glance when the worker nodes didn't
join the cluster.