upi/install/gcp: Create a JUnit breaking down install progress
Make it easier for TestGrid and other JUnit-based consumers to
distinguish between various provider-side failures and product-side
failures.  Although there may be mixed cases where some in-cluster
component has trouble with the provider but it ends up being
classified as a wait-for-bootstrap-complete test-case failure or
whatever.  It would also be nice to capture standard error output from
failing gcloud commands and whatnot and pipe that into the
JUnit-writer, but I haven't bothered with that while we collect
feedback on the initial sketch.