Fix boskos leasing to use info from job config instead of manually specifying it in the template.
Originally, our intent with boskos was to keep all of the identifying information as part of a config file that was tracked with the lease. Instead, since the information in question was needed to be available in the pod definition for DNS purposes, we instead focused on passing it along through the job definitions. This means that we can revert our boskos workflow to the standard leasing model.