Switch python print statements to be python2/python3 compatible
We use a number of `print ` statements in the repo, both in jobs and
libraries. This is valid python2, but not python3. Given python2 went
EOL months ago we need to catch up.

This should add the header which makes python2 act like python3 in
regards to print() statement and which is a noop after we switch.