ci-secret-generator: update registry values to be simple
The auth field in a .dockerconfigjson file is a basic-auth encoded pair
of username and password. The username in this case is entirely not
necessary as the password is a k8s authentication token and fully
identifies the user. When contacting the internal OpenShift registry,
the service account's k8s auth token is used and identification is
delegated, so we can just read the value of the token and construct our
own basic auth encoded string.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>