core-services/prow/02_config/_config: Highlight Boskos lease failures
Make this more obvious when it's the root cause of a job's failure.
I'd reported the non-obvious error message was reported upstream [1],
but actually pivoting to a more obvious error message was not trivial,
and the upstream report has since rotted closed.

It would also be nice to have ci-operator wrap these lease failures in
a JUnit test-case, but until that happens, just highlighting the
build-log line is a simple stopgap.