ci-operator/platform-balance/step-jobs-by-platform: Consider unused workflows for balanceable-ness
Load all the workloads defined in the step registry, and consider them
when figuring out alternatives for platform-stripped workflows.  For
example, we are very close to having a purely GCP update config at the
moment:

  $ git grep -rh 'workflow: openshift-upgrade-.*-loki' ci-operator/config | sort | uniq -c
        8     workflow: openshift-upgrade-aws-loki
       71     workflow: openshift-upgrade-gcp-loki

and pivoting the remaining jobs to openshift-upgrade-gcp-loki should
not block us from balancing back to openshift-upgrade-aws-loki later
on.