Add step-registry for periodic upgrade job for dropping net_raw
We are diverging the capabilities to its own MC in a patch in 4.5
so adding a periodic job to make sure that the new_raw capability
is still enabled in the restricted scc if the capabilities MC exists.
And this should persist across future upgrades if the capabilities is
not deleted.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>