oVirt e2e: temporarily increase resources to avoid BZ#1874696
For around 2 weeks ovirt PRs are RED because of BZ#1874696
After talking with the network team we understand that it is under active development but it will take some time.
We don't want to keep being blocked on the issue anymore.

This PR will increase the resources for each job to match the conformance jobs resources, that means the amount of parallel PRs our CI will be able to support will be half
(4 PRs), this is still better than all RED PRs.

It's also important to note that there will still be around 15%-20% chance of hitting BZ#1874696 but at least its not 95% like the current situation.

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>