oVirt e2e: increase teardown pod timeout
On e2e upgrade jobs sometimes the job takes more than 3H to finish the test,
When this happens the cluster is torn down and the test pod just keeps running till it gets a timeout.

This temporary workaround increases the timeout to 4H, to minimize the chance of this happening.
This edge case will be handled more elegantly later on.

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>