Mandate e2e-metal-ipi on cluster-network-operator and ovn-kubernetes
Multiple times now, pull requests have merged with e2e-metal-ipi failing
with explicitly clear messages in the CI summary that the failure was
CNO and/or OVN related. We are working on renaming these jobs starting
with 4.7 to indicate they are ovn-ipv6, but it's been proven now that
optional jobs are meaningless.

This sets e2e-metal-ipi to be required to pass on both CNO and OVN
master. These jobs are now as stable as any other platform, so there's
no reason to leave them as optional. It's also the only IPv6 coverage
that exists in OpenShift today.