Account for and mirror the release-controller secrets
The release-controller runs a couple of tools and also fires off jobs that each
require their own secrets.  Only 1 of the 3 secrets was accounted for in the
ci-secret-bootstrap, so I have added the other 2 and also added their respective
creation on the app.ci cluster (in preparation for relocating these pieces there).
I have also took care of the mirroring of all 3 of these secrets into the architecture
specific namespaces that run their own versions of the release-controller.