release-controller: update tools image and throttle github
This PR sets a github throttle to the release-controller deployments and
updates the github cache image to use 4.6, which has the required oc
version needed for the bugzilla functionality to properly function.

This also changes the `previousReleaseTag` versions for all releases to
a more reasonable release in order to reduce the number of requests to
bugzilla and github.