vSphere: remove the use of upi-installer image; which will prepare for upgrade testing
The `upi-installer` image was used in IPI because our previous
vSphere enviornment used a self-signed certificate.
Since moving to VMC it is a valid certificate we no longer need
to add it to the trust.

The `upi-installer` image also causes issues with trying
to test upgrades between versions since its not apart of
`release`.

Add the ability to install the `awscli` package via pip for
both python2 and python3 since 4.5 and 4.6 use different versions.