core-services/prow/02_config/_plugins: Allow ON_QA 4.7 for 4.6
Correcting 37697539c2 ([4.6 cut] Bugzilla criteria, 2020-10-02, #12400),
which allowed the 4.7 bug to be MODIFIED or VERIFIED, but not the
intermediate ON_QA.  This state is intended to allow teams to backport
to 4.6 once they've landed a 4.7 change.  Requiring a 4.7 change
reduces the risk of regressing in updates from a patched 4.6 to an
unpatched 4.7.  Not requiring VERIFIED or later allows QE to focus on
verifying 4.6 in the run-up to 4.6.0.  Once QE gets some spare
capacity and/or we expect no further 4.6.0 changes, we will pivot to
only allowing VERIFIED or later 4.7 bugs to unblock 4.6 backports.

Compare with 1a741c0e38 (Allow dependent bugs to be MODIFIED in 4.4,
2020-02-19, #7240).