core-services/prow/02_config: Enumerate multi-count names
The names are used for custom resources, so we need unique names for
each entry [1].  Consuming jobs will have to split on the --, which is
hopefully unique enough, to get the raw name (if they plan to use it
to select a region or similar).  I'm not using '-', because:

a. I wanted to avoid enumerating single-count entries.
b. I wanted consumers to not have to care about whether they were
   drawing from a single-count or multi-count entry.

Using a separator that is not part of the raw names means that
consumers can always split on that separator without having to worry
about whether they were drawing from a single-count or a multi-count
name.

[1]: https://github.com/openshift/release/pull/12589#discussion_r502457670