apply normal release rules to cluster-baremetal-operator
Work started in openshift/cluster-baremetal-operator while the 4.6
release was being finalized and the bugzilla rules were applied to
most repositories. Since the new component was not being included in
4.6, there was no reason to apply those rules. Now that 4.7 is opened
for development, we should remove the special treatment so the
repository is subject to the normal approval policies for the 4.7
release.

This change reverts 77f1d12f903bdfa88de8358e7adb8d29c60c30c2

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>