Remove redundant rbac manifests
We apply both the manifests in core-services and in
clusters/build-clusters/common. The latter includes manifests for both
the prow-controller-manager and hook, which means that they race and we
could end up with one or the other, resulting in hard to understand rbac
issues, because they might be fixed by the time someone looks at them.