Update release-controller-kubeconfigs with app.ci configurations
I am in the process of switching over a couple of the tools, that the
release-controller runs, from the api.ci cluster, over to the app.ci
cluster.  In order to do so, I need access to the kubeconfigs for the
app.ci cluster. This PR simply added the app.ci kubeconfigs, for the
various release-controller service accounts, to the secrets consumed
by the release-controller.