ci-operator/step-registry: Unify OPENSHIFT_*_RELEASE_IMAGE dependencies
Tests generally install one OpenShift release, and then perform some
actions to validate that release.  Sometimes the actions include
updating to a different release, or a series of different releases.
With this commit, I'm pivoting to dependency naming that more clearly
reflects these goals, instead of "RELEASE_IMAGE_LATEST", which was a
reflection of the default release:latest value, and not an expression
of the intended semantics.

I've shifted to explicitly declaring these image dependencies, taking
advantage of openshift/ci-tools@782008c873 (ci-operator: allow users
to depend on images in steps, 2020-07-30, openshift/ci-tools#1044),
instead of relying on the implicit RELEASE_IMAGE_LATEST and similar.

I've dropped explicit dependencies that were not used in a step's
associated commands, like RELEASE_IMAGE_LATEST in ipi-install since
eb8eb3239f (step-registry: add upgrade workflows, 2020-08-25, #11247).

I've dropped explicit dependency overrides that set the same value as
the default.  For example, there was no need to set:

  OPENSHIFT_UPGRADE_RELEASE_IMAGE: "release:latest"

in openshift-upgrade-aws-loki, because release:latest is already the
default in all steps that consume OPENSHIFT_UPGRADE_RELEASE_IMAGE.

I've dropped the _OVERRIDE suffix from the dependencies, because while
we may set that suffix when communicating with the openshift-install
command, it's a legitimate knob and not a strange override in each
steps' public API.